Fixed autoplay video and swaped the import in app.jsx to the home2 file
